added changes to code according to new code format
