+++ src/components/index.tsx
@@ -1,0 +1,2 @@
+export { default as App } from './app';
+export { default as Counter } from './counter';
