+++ src/react-app-env.d.ts
@@ -1,1 +1,1 @@
-/// <reference types="react-scripts" />
+/// <reference types='react-scripts' />
